+++ cartridge-cli.lua
@@ -1766,16 +1766,14 @@ local function pack_tgz()
     if not ok then return false, err end
 
     info('Create archive')
-    local data, err = check_output(
-        "cd %s && %s -cvzf - %s",
-        app_state.appfiles_dir, tar, app_state.name
+
+    local ok, err = call(
+        "cd %s && %s -czf %s %s",
+        app_state.appfiles_dir, tar, tgz_filepath, app_state.name
     )
-    if data == nil then
+    if not ok then
         return false, string.format("Failed to pack tgz: %s", err)
     end
-
-    local ok, err = utils.write_file(tgz_filepath, data)
-    if not ok then return false, err end
 
     info("Resulting tar.gz saved as: %s", tgz_filepath)
 
@@ -2103,6 +2101,46 @@ local function filter_out_known_files(files)
     return result
 end
 
+local function file_md5_hex(fullpath)
+    if not fio.path.exists(fullpath) then
+        return nil, string.format('File does not exist: %s', fullpath)
+    end
+
+    local md5sum = which('md5sum')
+
+    if md5sum == nil then
+        return nil, 'md5sum binary not found'
+    end
+
+    local output = check_output('%s %s', md5sum, fullpath)
+    if output == nil then
+        return nil, string.format('Failed to get %s MD5 using md5sum', fullpath)
+    end
+
+    local md5_hex = output:strip():split(' ', 1)[1]
+    return md5_hex
+end
+
+local function file_sha256_hex(fullpath)
+    if not fio.path.exists(fullpath) then
+        return nil, string.format('File does not exist: %s', fullpath)
+    end
+
+    local sha256sum = which('sha256sum')
+
+    if sha256sum == nil then
+        return nil, 'sha256sum binary not found'
+    end
+
+    local output = check_output('%s %s', sha256sum, fullpath)
+    if output == nil then
+        return nil, string.format('Failed to get %s SHA256 using sha256sum', fullpath)
+    end
+
+    local sha256_hex = output:strip():split(' ', 1)[1]
+    return sha256_hex
+end
+
 local function generate_fileinfo(source_dir)
     local function gen_dirnames(files)
         local dirnames = {}
@@ -2165,7 +2203,7 @@ local function generate_fileinfo(source_dir)
             table.insert(result.fileflags, 0)
             table.insert(result.filedigests, '')
         else
-            local filedigest, err = utils.file_md5_hex(fullpath)
+            local filedigest, err = file_md5_hex(fullpath)
             if filedigest == nil then return false, err end
 
             table.insert(result.fileflags, bit.lshift(1, 4))
@@ -2206,6 +2244,12 @@ local function pack_cpio(opts)
         return nil, "gzip binary is required to build rpm packages"
     end
 
+    local md5sum = which('md5sum')
+
+    if md5sum == nil then
+        return nil, "md5sum binary is required to build rpm packages"
+    end
+
     local distribution_dir = fio.pathjoin(app_state.appfiles_dir, '/usr/share/tarantool/', app_state.name)
     local ok, err = form_distribution_dir(distribution_dir)
     if not ok then return nil, err end
@@ -2219,32 +2263,38 @@ local function pack_cpio(opts)
     local files = utils.find_files(app_state.appfiles_dir, {include_dirs=true, exclude={'.git'}})
     files = filter_out_known_files(files)
 
-    local ok, err = utils.write_file(fio.pathjoin(app_state.appfiles_dir, 'files'), table.concat(files, '\n'))
+    local files_list_path = fio.pathjoin(app_state.build_dir, 'files')
+    local ok, err = utils.write_file(files_list_path, table.concat(files, '\n'))
     if not ok then return nil, err end
 
+    local raw_cpio_path = fio.pathjoin(app_state.build_dir, 'raw_cpio')
+
     info('Create CPIO archive')
-    local ok, pack_err = call("cd %s && cat files | %s -o -H newc > unpacked", app_state.appfiles_dir, cpio)
+    local ok, pack_err = call(
+        "cd %s && cat %s | %s -o -H newc > %s",
+        app_state.appfiles_dir,
+        files_list_path,
+        cpio,
+        raw_cpio_path
+    )
     if not ok then
         return nil, string.format("Failed to pack CPIO: %s", pack_err)
     end
 
-    info('Compress it using GZIP')
-    local payloadsize = fio.stat(fio.pathjoin(app_state.appfiles_dir, 'unpacked')).size
-    local archive, read_err = check_output("cd %s && cat unpacked | %s -9", app_state.appfiles_dir, gzip)
-    if archive == nil then
-        return nil, string.format("Failed to pack CPIO: %s", read_err)
-    end
+    local payloadsize = fio.stat(raw_cpio_path).size
 
-    for _, f in ipairs({'unpacked', 'files'}) do
-        local filepath = fio.pathjoin(app_state.appfiles_dir, f)
-        local ok, err = utils.remove_by_path(filepath)
-        if not ok then return nil, err end
+    info('Compress it using GZIP')
+
+    local archive_path = fio.pathjoin(app_state.build_dir, string.format('%s.cpio', app_state.name))
+    local ok, err = call("cat %s | %s -9 > %s", raw_cpio_path, gzip, archive_path)
+    if not ok then
+        return nil, string.format('Failed to create CPIO: %s', err)
     end
 
     local fileinfo = generate_fileinfo(app_state.appfiles_dir)
 
     return {
-        archive = archive,
+        path = archive_path,
         fileinfo = fileinfo,
         payloadsize = payloadsize,
     }
@@ -2272,7 +2322,7 @@ local function pack_rpm(opts)
     info('Construct RPM header')
     -- compute payload digest
     local payloaddigest_algo = PGPHASHALGO_SHA256
-    local payloaddigest = digest.sha256_hex(cpio.archive)
+    local payloaddigest = file_sha256_hex(cpio.path)
 
     local create_user_script_rpm = CREATE_USER_SCRIPT
 
@@ -2339,10 +2389,18 @@ local function pack_rpm(opts)
         HEADERIMMUTABLE
     )
 
-    local body = header .. cpio.archive
-    local md5 = digest.md5(body)
+    local body_filepath = fio.pathjoin(app_state.build_dir, 'body')
+    utils.write_file(body_filepath, header)
+
+    local ok, err = call('cat %s >> %s', cpio.path, body_filepath)
+    if not ok then
+        return nil, string.format('Failed to write RPM archive body: %s', err)
+    end
+
+    local md5 = string.fromhex(file_md5_hex(body_filepath))
+    local sig_size = fio.stat(body_filepath).size
     local sha1 = digest.sha1_hex(header)
-    local sig_size = #body
+
     local signature_header = gen_header(
         {
             {'SHA1', 'STRING', sha1},
@@ -2354,11 +2412,16 @@ local function pack_rpm(opts)
         HEADERSIGNATURES
     )
 
-    body = lead .. utils.buf_pad_to_8_byte_boundary(signature_header) .. body
-
     info('Write RPM file')
-    local ok, err = utils.write_file(rpm_filepath, body)
-    if not ok then return false, err end
+
+    local rpm_header = lead .. utils.buf_pad_to_8_byte_boundary(signature_header)
+    local ok, err = utils.write_file(rpm_filepath, rpm_header)
+    if not ok then return nil, err end
+
+    local ok, err = call('cat %s >> %s', body_filepath, rpm_filepath)
+    if not ok then
+        return nil, string.format('Failed to append archive body to RPM: %s', err)
+    end
 
     info("Resulting rpm saved as: %s", rpm_filepath)
 
@@ -2460,12 +2523,10 @@ local function pack_deb(opts)
     if not ok then return false, err end
 
     info('Archive deb control data')
-    local control_data, pack_control_err = check_output("cd %s && %s -cvJf - .", control_dir, tar)
-    if control_data == nil then
-        die('Failed to pack deb control files: %s', pack_control_err)
+    local ok, err = call("cd %s && %s -cJf %s .", control_dir, tar, control_tgz_path)
+    if not ok  then
+        return false, string.format('Failed to pack deb control files: %s', err)
     end
-    local ok, err = utils.write_file(control_tgz_path, control_data)
-    if not ok then return false, err end
 
     -- data.tar.xz
     info('Generate package data')
@@ -2485,12 +2546,10 @@ local function pack_deb(opts)
     if not ok then return false, err end
 
     info('Compress package data using TAR')
-    local data, pack_data_err = check_output("cd %s && %s -cvJf - .", data_dir, tar)
-    if data == nil then
-        die('Failed to pack deb package files: %s', pack_data_err)
+    local ok, err = call("cd %s && %s -cJf %s .", data_dir, tar, data_tgz_path)
+    if not ok then
+        return nil, string.format('Failed to pack deb package files: %s', err)
     end
-    local ok, err = utils.write_file(data_tgz_path, data)
-    if not ok then return false, err end
 
     -- pack .deb
     info('Pack DEB archive')
@@ -2504,7 +2563,7 @@ local function pack_deb(opts)
         app_state.appfiles_dir, ar, deb_filename, archive_files
     )
     if not ok then
-        die('Failed to pack DEB package: %s', pack_deb_err)
+        return nil, string.format('Failed to pack DEB package: %s', pack_deb_err)
     end
 
     local ok, err = utils.copyfile(fio.pathjoin(app_state.appfiles_dir, deb_filename), app_state.dest_dir)